python/requirement-gpu.txt
appdirs==1.4.0
boto==2.46.1
bz2file==0.98
funcsigs==1.0.2
gensim==1.0.0
h5py==2.6.0
Keras==1.2.2
mock==2.0.0
numpy==1.12.0
packaging==16.8
pbr==1.10.0
protobuf==3.2.0
pyparsing==2.1.10
PyYAML==3.12
requests==2.13.0
scikit-learn==0.18.1
scipy==0.18.1
six==1.10.0
smart-open==1.4.0
tensorflow-gpu==1.0.0
Theano==0.8.2
